DEV: refactored BaseMultiImage classes and module
